requirements.txt
asgiref==3.7.2
astroid==3.0.2
black==23.12.1
click==8.1.7
colorama==0.4.6
dill==0.3.7
Django==5.0.1
django-ckeditor==6.7.0
django-cors-headers==4.3.1
django-js-asset==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
Faker==22.0.0
flake8==7.0.0
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
pycodestyle==2.11.1
pyflakes==3.2.0
PyJWT==2.8.0
pylint==3.0.3
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.12.3
typing_extensions==4.9.0
tzdata==2023.4
